[PATCH] I2C: Spare 1 byte in lm90 driver
I just noticed the other day that the lm90 driver uses an u16 to store
the value of the 8-bit alarms register. This is most probably due to the
fact that I originally copied the lm90 driver from the lm83 driver,
which actually has two 8-bit registers for alarms, and obviously forgot
to change the variable type.

+++ drivers/i2c/chips/lm90.c
@@ -185,7 +185,7 @@ struct lm90_data {
 	s16 temp_input2, temp_low2, temp_high2; /* remote, combined */
 	s8 temp_crit1, temp_crit2;
 	u8 temp_hyst;
-	u16 alarms; /* bitvector, combined */
+	u8 alarms; /* bitvector */
 };
 
 /*
